feat(rm): move device removal from realm management to realm management api
move device removal
move accessory query functions
move tests and generators
fixed references
