+++ apps/astarte_realm_management/test/test_helper.exs
@@ -17,7 +17,6 @@
 #
 
 Mimic.copy(Astarte.DataAccess.Config)
-Mimic.copy(Astarte.RealmManagement.DeviceRemoval.Core)
 Mimic.copy(Task.Supervisor)
 
 ExUnit.start(capture_log: true)
