+++ apps/astarte_realm_management_api/test/support/generators/deletion_in_progress.ex
@@ -18,7 +18,7 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-defmodule Astarte.RealmManagement.Generators.DeletionInProgress do
+defmodule Astarte.RealmManagement.API.Generators.DeletionInProgress do
   @moduledoc """
   Generator for `Astarte.DataAccess.Device.DeletionInProgress`
   """
