+++ apps/astarte_realm_management_api/test/support/generators/individual_property.ex
@@ -18,7 +18,7 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-defmodule Astarte.RealmManagement.Generators.IndividualProperty do
+defmodule Astarte.RealmManagement.API.Generators.IndividualProperty do
   @moduledoc """
   Generators of `IndividualProperty`es
   """
